* ext/openssl/ossl_bn.c (ossl_bn_s_rand, ossl_bn_s_pseudo_rand):   Int should be enough here.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_8@16692 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/openssl/ossl_rand.c
@@ -96,7 +96,7 @@ static VALUE
 ossl_rand_bytes(VALUE self, VALUE len)
 {
     VALUE str;
-    long n = NUM2INT(len);
+    int n = NUM2INT(len);
 
     str = rb_str_new(0, n);
     if (!RAND_bytes(RSTRING_PTR(str), n)) {
@@ -115,7 +115,7 @@ static VALUE
 ossl_rand_pseudo_bytes(VALUE self, VALUE len)
 {
     VALUE str;
-    long n = NUM2INT(len);
+    int n = NUM2INT(len);
 
     str = rb_str_new(0, n);
     if (!RAND_pseudo_bytes(RSTRING_PTR(str), n)) {
